React Select Bundle Improvement
