+++ packages/app-store/routing-forms/components/react-awesome-query-builder/widgets.tsx
@@ -1,3 +1,4 @@
+import dynamic from "next/dynamic";
 import type { ChangeEvent } from "react";
 import type {
   ButtonGroupProps,
@@ -7,8 +8,12 @@ import type {
   ProviderProps,
 } from "react-awesome-query-builder";
 
-import { Button as CalButton, SelectWithValidation as Select, TextField, TextArea } from "@calcom/ui";
+import { Button as CalButton, TextField, TextArea } from "@calcom/ui";
 import { Trash, Plus } from "@calcom/ui/components/icon";
+
+const Select = dynamic(
+  async () => (await import("@calcom/ui")).SelectWithValidation
+) as unknown as typeof import("@calcom/ui").SelectWithValidation;
 
 export type CommonProps<
   TVal extends
